# Optimize hash map capacity calculations